refactor: add add_project_handler module instead of handling `addproject` command directly in main module
